// sdk/src/gateway/generated-client/models/GatewayQuoteOneOf.ts
/* tslint:disable */
/* eslint-disable */
/**
 * gateway-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { GatewayOnrampQuote } from './GatewayOnrampQuote';
import {
    GatewayOnrampQuoteFromJSON,
    GatewayOnrampQuoteFromJSONTyped,
    GatewayOnrampQuoteToJSON,
    GatewayOnrampQuoteToJSONTyped,
} from './GatewayOnrampQuote';

/**
 *
 * @export
 * @interface GatewayQuoteOneOf
 */
export interface GatewayQuoteOneOf {
    /**
     *
     * @type {GatewayOnrampQuote}
     * @memberof GatewayQuoteOneOf
     */
    onramp: GatewayOnrampQuote;
}

/**
 * Check if a given object implements the GatewayQuoteOneOf interface.
 */
export function instanceOfGatewayQuoteOneOf(value: object): value is GatewayQuoteOneOf {
    if (!('onramp' in value) || value['onramp'] === undefined) return false;
    return true;
}

export function GatewayQuoteOneOfFromJSON(json: any): GatewayQuoteOneOf {
    return GatewayQuoteOneOfFromJSONTyped(json, false);
}

export function GatewayQuoteOneOfFromJSONTyped(json: any, ignoreDiscriminator: boolean): GatewayQuoteOneOf {
    if (json == null) {
        return json;
    }
    return {
        onramp: GatewayOnrampQuoteFromJSON(json['onramp']),
    };
}

export function GatewayQuoteOneOfToJSON(json: any): GatewayQuoteOneOf {
    return GatewayQuoteOneOfToJSONTyped(json, false);
}

export function GatewayQuoteOneOfToJSONTyped(
    value?: GatewayQuoteOneOf | null,
    ignoreDiscriminator: boolean = false
): any {
    if (value == null) {
        return value;
    }

    return {
        onramp: GatewayOnrampQuoteToJSON(value['onramp']),
    };
}