// sdk/src/gateway/generated-client/models/GatewayQuoteOneOf1.ts
/* tslint:disable */
/* eslint-disable */
/**
 * gateway-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { GatewayOfframpQuote } from './GatewayOfframpQuote';
import {
    GatewayOfframpQuoteFromJSON,
    GatewayOfframpQuoteFromJSONTyped,
    GatewayOfframpQuoteToJSON,
    GatewayOfframpQuoteToJSONTyped,
} from './GatewayOfframpQuote';

/**
 *
 * @export
 * @interface GatewayQuoteOneOf1
 */
export interface GatewayQuoteOneOf1 {
    /**
     *
     * @type {GatewayOfframpQuote}
     * @memberof GatewayQuoteOneOf1
     */
    offramp: GatewayOfframpQuote;
}

/**
 * Check if a given object implements the GatewayQuoteOneOf1 interface.
 */
export function instanceOfGatewayQuoteOneOf1(value: object): value is GatewayQuoteOneOf1 {
    if (!('offramp' in value) || value['offramp'] === undefined) return false;
    return true;
}

export function GatewayQuoteOneOf1FromJSON(json: any): GatewayQuoteOneOf1 {
    return GatewayQuoteOneOf1FromJSONTyped(json, false);
}

export function GatewayQuoteOneOf1FromJSONTyped(json: any, ignoreDiscriminator: boolean): GatewayQuoteOneOf1 {
    if (json == null) {
        return json;
    }
    return {
        offramp: GatewayOfframpQuoteFromJSON(json['offramp']),
    };
}

export function GatewayQuoteOneOf1ToJSON(json: any): GatewayQuoteOneOf1 {
    return GatewayQuoteOneOf1ToJSONTyped(json, false);
}

export function GatewayQuoteOneOf1ToJSONTyped(
    value?: GatewayQuoteOneOf1 | null,
    ignoreDiscriminator: boolean = false
): any {
    if (value == null) {
        return value;
    }

    return {
        offramp: GatewayOfframpQuoteToJSON(value['offramp']),
    };
}