// sdk/src/gateway/generated-client/models/GatewayQuoteOneOf2.ts
/* tslint:disable */
/* eslint-disable */
/**
 * gateway-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { GatewayLayerZeroQuote } from './GatewayLayerZeroQuote';
import {
    GatewayLayerZeroQuoteFromJSON,
    GatewayLayerZeroQuoteFromJSONTyped,
    GatewayLayerZeroQuoteToJSON,
    GatewayLayerZeroQuoteToJSONTyped,
} from './GatewayLayerZeroQuote';

/**
 *
 * @export
 * @interface GatewayQuoteOneOf2
 */
export interface GatewayQuoteOneOf2 {
    /**
     *
     * @type {GatewayLayerZeroQuote}
     * @memberof GatewayQuoteOneOf2
     */
    layerZero: GatewayLayerZeroQuote;
}

/**
 * Check if a given object implements the GatewayQuoteOneOf2 interface.
 */
export function instanceOfGatewayQuoteOneOf2(value: object): value is GatewayQuoteOneOf2 {
    if (!('layerZero' in value) || value['layerZero'] === undefined) return false;
    return true;
}

export function GatewayQuoteOneOf2FromJSON(json: any): GatewayQuoteOneOf2 {
    return GatewayQuoteOneOf2FromJSONTyped(json, false);
}

export function GatewayQuoteOneOf2FromJSONTyped(json: any, ignoreDiscriminator: boolean): GatewayQuoteOneOf2 {
    if (json == null) {
        return json;
    }
    return {
        layerZero: GatewayLayerZeroQuoteFromJSON(json['layerZero']),
    };
}

export function GatewayQuoteOneOf2ToJSON(json: any): GatewayQuoteOneOf2 {
    return GatewayQuoteOneOf2ToJSONTyped(json, false);
}

export function GatewayQuoteOneOf2ToJSONTyped(
    value?: GatewayQuoteOneOf2 | null,
    ignoreDiscriminator: boolean = false
): any {
    if (value == null) {
        return value;
    }

    return {
        layerZero: GatewayLayerZeroQuoteToJSON(value['layerZero']),
    };
}